Refactor code (use links/verbose modules)
Resolves: #24
